requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
psycopg[binary]>=3.1,<4.0
python-dotenv>=1.0,<2.0
django-cors-headers>=4.0,<5.0
django-extensions>=3.2,<4.0
